fix(grids) Grid Element Direct Lookup Ordered Correctly. - Moved a lot of common types and functions used when obtaining values   from elements for grids into `grid-measurements.tsx`. - Added `getGridElementMeasurementHelperData` to retrieve the   necessary details for a grid element, along with some supporting functions. - Extracted out a `getGlobalFrame` function for the common lookup used. - Changed some test ID values in two components as they created the same ones. - `PaddingResizeControl` set to the `bottom` priority so that it doesn't end up   on top of the track resizing controls. - `GridControl` now gets the cell dimensions directly from the grid element. - `GridElementContainingBlock` gets the grid child dimensions directly. - Added `HelperControlsStateContext` and made the connected change to `StoreName`. - Added `helperControlsStore` to `Editor`. - Added contexts for the helper store to `EditorRoot` and `HotRoot`. - Wrapped the contents of `GridElementContainingBlocks` and `GridMeasurementHelpers`   with the new helper store context. - Added calls to update the helper store in the regular and test dispatch flows.
