+++ editor/src/templates/editor.tsx
@@ -58,6 +58,7 @@ import {
   CanvasStateContext,
   createStoresAndState,
   EditorStateContext,
+  HelperControlsStateContext,
   LowPriorityStateContext,
   OriginalMainEditorStateContext,
 } from '../components/editor/store/store-hook'
@@ -221,6 +222,7 @@ export class Editor {
   storedState: EditorStoreFull
   utopiaStoreHook: UtopiaStoreAPI
   canvasStore: UtopiaStoreAPI
+  helperControlsStore: UtopiaStoreAPI
   lowPriorityStore: UtopiaStoreAPI
   spyCollector: UiJsxCanvasContextData = emptyUiJsxCanvasContextData()
   domWalkerMutableState: DomWalkerMutableStateData
@@ -247,6 +249,7 @@ export class Editor {
         this.boundDispatch,
         this.utopiaStoreHook,
         this.canvasStore,
+        this.helperControlsStore,
         this.lowPriorityStore,
         this.spyCollector,
         this.domWalkerMutableState,
@@ -298,6 +301,10 @@ export class Editor {
       patchedStoreFromFullStore(this.storedState, 'canvas-store'),
     )
 
+    const helperControlsStore = createStoresAndState(
+      patchedStoreFromFullStore(this.storedState, 'helper-controls-store'),
+    )
+
     const lowPriorityStore = createStoresAndState(
       patchedStoreFromFullStore(this.storedState, 'low-priority-store'),
     )
@@ -305,6 +312,8 @@ export class Editor {
     this.utopiaStoreHook = store
 
     this.canvasStore = canvasStore
+
+    this.helperControlsStore = helperControlsStore
 
     this.lowPriorityStore = lowPriorityStore
 
@@ -490,6 +499,9 @@ export class Editor {
           ReactDOM.flushSync(() => {
             ReactDOM.unstable_batchedUpdates(() => {
               this.canvasStore.setState(patchedStoreFromFullStore(this.storedState, 'canvas-store'))
+              this.helperControlsStore.setState(
+                patchedStoreFromFullStore(this.storedState, 'helper-controls-store'),
+              )
             })
           })
         })
@@ -727,6 +739,7 @@ export const EditorRoot: React.FunctionComponent<{
   dispatch: EditorDispatch
   mainStore: UtopiaStoreAPI
   canvasStore: UtopiaStoreAPI
+  helperControlsStore: UtopiaStoreAPI
   lowPriorityStore: UtopiaStoreAPI
   spyCollector: UiJsxCanvasContextData
   domWalkerMutableState: DomWalkerMutableStateData
@@ -734,6 +747,7 @@ export const EditorRoot: React.FunctionComponent<{
   dispatch,
   mainStore,
   canvasStore,
+  helperControlsStore,
   lowPriorityStore,
   spyCollector,
   domWalkerMutableState,
@@ -748,15 +762,17 @@ export const EditorRoot: React.FunctionComponent<{
             <EditorStateContext.Provider value={mainStore}>
               <DomWalkerMutableStateCtx.Provider value={domWalkerMutableState}>
                 <CanvasStateContext.Provider value={canvasStore}>
-                  <LowPriorityStateContext.Provider value={lowPriorityStore}>
-                    <UiJsxCanvasCtxAtom.Provider value={spyCollector}>
-                      <AnimationContext.Provider
-                        value={{ animate: animate, scope: animationScope }}
-                      >
-                        <EditorComponent />
-                      </AnimationContext.Provider>
-                    </UiJsxCanvasCtxAtom.Provider>
-                  </LowPriorityStateContext.Provider>
+                  <HelperControlsStateContext.Provider value={helperControlsStore}>
+                    <LowPriorityStateContext.Provider value={lowPriorityStore}>
+                      <UiJsxCanvasCtxAtom.Provider value={spyCollector}>
+                        <AnimationContext.Provider
+                          value={{ animate: animate, scope: animationScope }}
+                        >
+                          <EditorComponent />
+                        </AnimationContext.Provider>
+                      </UiJsxCanvasCtxAtom.Provider>
+                    </LowPriorityStateContext.Provider>
+                  </HelperControlsStateContext.Provider>
                 </CanvasStateContext.Provider>
               </DomWalkerMutableStateCtx.Provider>
             </EditorStateContext.Provider>
@@ -773,17 +789,27 @@ export const HotRoot: React.FunctionComponent<{
   dispatch: EditorDispatch
   mainStore: UtopiaStoreAPI
   canvasStore: UtopiaStoreAPI
+  helperControlsStore: UtopiaStoreAPI
   lowPriorityStore: UtopiaStoreAPI
   spyCollector: UiJsxCanvasContextData
   domWalkerMutableState: DomWalkerMutableStateData
 }> = hot(
-  ({ dispatch, mainStore, canvasStore, lowPriorityStore, spyCollector, domWalkerMutableState }) => {
+  ({
+    dispatch,
+    mainStore,
+    canvasStore,
+    helperControlsStore,
+    lowPriorityStore,
+    spyCollector,
+    domWalkerMutableState,
+  }) => {
     return (
       <EditorRoot
         dispatch={dispatch}
         spyCollector={spyCollector}
         mainStore={mainStore}
         canvasStore={canvasStore}
+        helperControlsStore={helperControlsStore}
         lowPriorityStore={lowPriorityStore}
         domWalkerMutableState={domWalkerMutableState}
       />
@@ -796,6 +822,7 @@ async function renderRootComponent(
   dispatch: EditorDispatch,
   mainStore: UtopiaStoreAPI,
   canvasStore: UtopiaStoreAPI,
+  helperControlsStore: UtopiaStoreAPI,
   lowPriorityStore: UtopiaStoreAPI,
   spyCollector: UiJsxCanvasContextData,
   domWalkerMutableState: DomWalkerMutableStateData,
@@ -813,6 +840,7 @@ async function renderRootComponent(
             mainStore={mainStore}
             spyCollector={spyCollector}
             canvasStore={canvasStore}
+            helperControlsStore={helperControlsStore}
             lowPriorityStore={lowPriorityStore}
             domWalkerMutableState={domWalkerMutableState}
           />,
@@ -824,6 +852,7 @@ async function renderRootComponent(
             spyCollector={spyCollector}
             mainStore={mainStore}
             canvasStore={canvasStore}
+            helperControlsStore={helperControlsStore}
             lowPriorityStore={lowPriorityStore}
             domWalkerMutableState={domWalkerMutableState}
           />,
